able to scan barcode and get back json info on home screen
