+++ app/build.gradle
@@ -60,7 +60,6 @@ dependencies {
     implementation 'com.google.firebase:firebase-auth:21.1.0'
     implementation 'com.google.android.gms:play-services-auth:20.3.0'
     implementation 'com.github.bumptech.glide:glide:4.11.0'
-    apply plugin: 'com.google.gms.google-services'
 
     // Import the BoM for the Firebase platform
     implementation platform('com.google.firebase:firebase-bom:31.1.0')
@@ -83,3 +82,4 @@ dependencies {
     implementation("com.squareup.okhttp3:okhttp:3.12.10")
     implementation("com.squareup.okhttp3:logging-interceptor")
 }
+apply plugin: 'com.google.gms.google-services'
