+++ app/src/main/AndroidManifest.xml
@@ -53,9 +53,6 @@
             android:screenOrientation="portrait"
             android:stateNotNeeded="true"
             android:theme="@style/zxing_CaptureTheme"/>
-        <activity
-            android:name=".JsonResult"
-            android:parentActivityName=".MainActivity" />
     </application>
 
 </manifest>
